CH3. PP.5
Write a program that allows the user to convert a temperature given in
degrees from either
Celsius to Fahrenheit or Fahrenheit to Celsius. Prompt the user to enter
a temperature and either
a C or c for Celsius or an F or F for Fahrenheit. Convert the
temperature to the opposite one.
Display the result in a readable format. If anything other than C,c,F,
or f is entered,
print an error message and stop.